+++ src/include/setup.hpp
@@ -21,7 +21,7 @@ class Setup {
 
   static const std::string& getMinersConfig();
 
-  static const std::string getMinedBy();
+  static const std::string getSubmittedBy();
 
   static void setMinersConfig(const std::string& loda_config);
 
@@ -70,7 +70,7 @@ class Setup {
   static bool checkMiningMode();
   static bool checkMinersConfig();
   static bool checkMineParallelScript();
-  static bool checkMinedBy();
+  static bool checkSubmittedBy();
   static bool checkMaxMemory();
 
   static bool updateFile(const std::string& local_file, const std::string& url,