+++ src/oeis_manager.cpp
@@ -473,7 +473,7 @@ void OeisManager::addSeqComments(Program &p) const {
 }
 
 void OeisManager::dumpProgram(size_t id, Program p, const std::string &file,
-                              const std::string &mined_by) const {
+                              const std::string &submitted_by) const {
   ProgramUtil::removeOps(p, Operation::Type::NOP);
   ProgramUtil::removeComments(p);
   addSeqComments(p);
@@ -481,8 +481,9 @@ void OeisManager::dumpProgram(size_t id, Program p, const std::string &file,
   std::ofstream out(file);
   auto &seq = sequences.at(id);
   out << "; " << seq << std::endl;
-  if (!mined_by.empty()) {
-    out << "; Mined by " << mined_by << std::endl;
+  if (!submitted_by.empty()) {
+    out << "; " << ProgramUtil::SUBMITTED_BY_PREFIX << " " << submitted_by
+        << std::endl;
   }
   out << "; " << seq.getTerms(OeisSequence::DEFAULT_SEQ_LENGTH) << std::endl;
   out << std::endl;
@@ -492,13 +493,13 @@ void OeisManager::dumpProgram(size_t id, Program p, const std::string &file,
 
 void OeisManager::alert(Program p, size_t id, const std::string &prefix,
                         const std::string &color,
-                        const std::string &mined_by) const {
+                        const std::string &submitted_by) const {
   auto &seq = sequences.at(id);
   std::stringstream buf;
   buf << prefix << " program for " << seq
       << " Terms: " << seq.getTerms(settings.num_terms);
-  if (!mined_by.empty()) {
-    buf << ". Mined by " << mined_by;
+  if (!submitted_by.empty()) {
+    buf << ". " << ProgramUtil::SUBMITTED_BY_PREFIX << " " << submitted_by;
   }
   auto msg = buf.str();
   Log::AlertDetails details;
@@ -596,7 +597,7 @@ std::pair<bool, bool> OeisManager::updateProgram(size_t id, const Program &p) {
   auto &seq = sequences.at(id);
   const std::string global_file = seq.getProgramPath(false);
   const std::string local_file = seq.getProgramPath(true);
-  const std::string mined_by = ProgramUtil::getMinedBy(p);
+  const std::string submitted_by = ProgramUtil::getSubmittedBy(p);
   bool is_new = true;
   std::string change;
 
@@ -633,15 +634,15 @@ std::pair<bool, bool> OeisManager::updateProgram(size_t id, const Program &p) {
 
   // write new or optimized program version
   if (Setup::getMiningMode() == MINING_MODE_SERVER) {
-    dumpProgram(id, minimized.second, global_file, mined_by);
+    dumpProgram(id, minimized.second, global_file, submitted_by);
   } else {
-    dumpProgram(id, minimized.second, local_file, mined_by);
+    dumpProgram(id, minimized.second, local_file, submitted_by);
   }
 
   // send alert
   std::string prefix = is_new ? "First" : change;
   std::string color = is_new ? "good" : "warning";
-  alert(minimized.second, id, prefix, color, mined_by);
+  alert(minimized.second, id, prefix, color, submitted_by);
 
   return {true, is_new};
 }